+++ packages/libra-rpc-graphql/src/index.ts
@@ -1,9 +1,5 @@
-import * as server from './server';
-
 export * from './client';
 export * from './graphql';
 export * from './link';
 export * from './resolvers';
 export * from './types';
-
-export {server};