+++ packages/libra-rpc-graphql/src/server/index.ts
@@ -1,6 +1,0 @@
-import * as graphql from './graphql';
-import * as variables from './variables';
-
-export * from './server';
-
-export {graphql, variables};