+++ packages/libra-rpc-graphql/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Changed
+
+- `server` is no longer available in this package, it can now be imported through `@shopify/libra-rpc-graphql-server` ([#25][https://github.com/shopify/libra-web-tools/pull/25])
 
 ## [2.5.2]